+++ NEWS.md
@@ -3,6 +3,7 @@
 ## trajr 1.3.0.9000
 
 * Allow the creation of trajectories with 0 points in `TrajFromCoords`.
+* Fixed bug in TrajSpeedIntervals: no intervals were being returned if the entire trajectory qualified as an interval.
 * Fix bug in plot.TrajSpeedIntervals: time (x-axis) didn't handle trajectories with a non-zero starting time.
 * Added functions to assist in simulating trajectories that are bounded or vary spatially in some way. The new functions are `TrajInPolygon`, `TrajSplit`, `TrajMerge` and `TrajSplitAtFirstCrossing`.
 * Added functions to calculate velocity and acceleration: `TrajVelocity` and `TrajAcceleration`. The documentation for `TrajDerivatives` has been updated to state that the `acceleration` returned is not really acceleration, rather it is change in speed over time.